Created the 2014-siliconvalley event.
Just has the basic announcement of the dates and also the Call For
Papers

I left the propose page the same except updated the dates for the call
for papers.

I emptied out all the sponsors (but left the sponsor page intact with a
call for sponsorship for each level except Media.

Deleted all the sponsor logos

I emptied out all the presentations and left just the header

The registration page just says it will be announced soon and added the
link to the mailing list to be kept informed.

Made the location page just say Silicon Valley TBD

Updated all the supporting _*.txt files so they have the proper dates,
email addresses, and other info.

Could not get it tested as I can't seem to get webby to run without an
error.

So you should make sure the pages look good.